+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+Refactor OneDocker CLI and OneDocker Service to support OneDocker E2E testing framework
 ### Removed
 
 ## [0.5.0] - 2023-3-21