+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add Insights Service
 ### Changed
-Refactor OneDocker CLI and OneDocker Service to support OneDocker E2E testing framework
+- Refactor OneDocker CLI and OneDocker Service to support OneDocker E2E testing framework
 ### Removed
 
 ## [0.5.0] - 2023-3-21