+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add LimitExceeded Exception as PcpError
 - ECS Gateway Run Task support for overriding task role
 - Container Service Create Instance support for Container permission configuration
+- OneDocker Service Start Containers support for Container permission configuration
 ### Changed
 - Refactor OneDocker CLI and OneDocker Service to support OneDocker E2E testing framework
 - Update Container Instance to include Container Permission, when configured